# OAS/index.yaml
openapi: 3.0.0
info:
  title: Pluto Interview Assistant API
  version: 1.0.0
  description: >
    REST API for Pluto Interview Assistant. All responses use a consistent
    envelope:

    { success: boolean, message: string|null, data: object|null, error:
    object|null }.
servers:
  - url: 'localhost:8080'
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/register:
    post:
      summary: Register user
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: randomemail@email.com
                password:
                  type: string
                  example: thisisapassword
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    nullable: true
                  data:
                    type: object
                    properties:
                      username:
                        type: string
                        example: "plutochan"
                      token:
                        type: string
                      expiresIn:
                        type: integer
                        example: 259200
                  error:
                    type: object
                    description: "Empty on success"
                    nullable: true
        '409':
          description: Email address already used
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Email address already used."
                  data:
                    type: object
                  error:
                    type: object
  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: randomemail@email.com
                password:
                  type: string
                  example: thisisapassword
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      expiresIn:
                        type: integer
                        example: 3600
                  error:
                    type: object
                    nullable: true
        '401':
          description: Wrong Password
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Wrong password
                  data:
                    type: object
                    nullable: true
                  error:
                    type: object
                    nullable: true
        '404':
            description: Email Not Registered
            content:
                application/json:
                    schema:
                        type: object
                        properties:
                            success:
                                type: boolean
                                example: false
                            message:
                                type: string
                                example: Email address not registered.
                            data:
                                type: object
                                nullable: true
                            error:
                                type: object
                                nullable: true
...
  /question-bank/questions:
    get:
      summary: Get questions (by conditions)
      tags:
        - "Question Bank"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                "offset": 
                  type: integer
                  description: "Default to 0"
                  nullable: true
                  example: 0
                "limit":
                  type: integer
                  description: "Default to 10"
                  nullable: true
                  example: 10
                "query":
                  type: string
                  description: Currently support title search only. Would add content search in future
                  example: "Algorithm"
                  nullable: true
                "questionType":
                  type: string
                  description: Must be in (Behavioural, Technical, Coding, System Design, Leadership, Product)
                  nullable: true
                "difficultyLevel":
                  type: string
                  description: Must be in (Easy, Medium, Hard)
                  nullable: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    nullable: true
                  data:
                    type: object
                    properties:
                    #   "questionType":
                    #     type: string
                    #     nullable: true
                    #   "difficultyLevel": 
                    #     type: string
                    #     nullable: true
                      questions:
                        type: array
                        items:
                          type: object
                          properties:
                            questionId:
                              type: string
                            title:
                              type: string
                              example: "Bubble Sort Algorithm"
                            # content:
                            #   type: string
                            questionType:
                              type: string
                              example: Coding
                            difficultyLevel:
                              type: string
                              example: Medium
                            "from":
                              type: string
                              nullable: true
                              example: "Microsoft"
                            # tags:
                            #   type: array
                            #   items:
                            #     type: string
                            #     nullable: true
                            #     example: "Algorithm"
                      pagination:
                        type: object
                        properties:
                            offset:
                                type: integer
                                example: 0  
                            limit:
                                type: integer
                                example: 10
                            pageCount:
                                type: integer
                                example: 2
                            total:
                                type: integer
                                example: 25
                            # isLastPage:
                            #     type: boolean
                            #     example: false
                  error:
                    type: object
                    nullable: true
    post:
        tags:
          - "Question Bank"
        summary: Add a New Question
        requestBody:
            content:
              application/json:
                schema:
                    type: object
                    properties:
                        title:
                            type: string
                            example: Explain Dependency Injection in Spring Boot
                        content:
                            type: string
                            description: No content is needed if it's a short question, 
                                e.g. the title is enough to explain the question
                            nullable: true
                        "questionType":
                            type: string
                            description: Must be in (Behavioural, Technical, Coding, System Design, Leadership, Product)
                            example: Technical
                        "difficultyLevel":
                            type: string
                            description: Must be in (Easy, Medium, Hard)
                            example: Medium
                        from:
                            type: string
                            description: The name of the company that has asked this question in interviews
                            nullable: true
                            example: Apple
        responses:
            '200':
                description: OK
                content:
                  application/json:
                    schema:
                        type: object
                        properties:
                            success:
                                type: boolean
                                example: true
                            message:
                                type: string
                                nullable: true
                            data:
                                type: object
                                nullable: true
                            error: 
                                type: object
                                nullable: true
            '400':
                description: Illegal Question Types or Difficulty Level
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                success:
                                    type: boolean
                                message:
                                    type: string
                                data:
                                    type: object
                                    nullable: true
                                error: 
                                    type: object
                                    nullable: true
                        examples:
                            illegalQuestionType:
                                summary: Invalid Question Type Error
                                value:
                                    success: false
                                    message: Illegal Question Type Provided
                                    data: null
                                    error: null
                            illegalDifficultyLevel:
                                summary: Invalid Difficulty Level Error
                                value:
                                    success: false # Changed from true
                                    message: Illegal Difficulty Level Provided
                                    data: null
                                    error: null
...